+++ src/main/java/sprint4/pom/SuccessfulOrderPage.java
@@ -1,4 +1,4 @@
-package Sprint_4.pom;
+package sprint4.pom;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.WebDriver;
@@ -7,7 +7,7 @@ public class SuccessfulOrderPage {
     //Окно успешного оформления заказа
 
     //Сообщение "Заказ оформлен"
-    private final By successMessage = By.xpath("//*[@class='Order_ModalHeader__3FDaJ' and contains(text(), 'Заказ оформлен')]");
+    public final By successMessage = By.xpath("//*[@class='Order_ModalHeader__3FDaJ' and contains(text(), 'Заказ оформлен')]");
 
     private final WebDriver webDriver;
